Perf: Improve Database Creation Speed
The create_db.py file can now create the database much faster. This
is a result of no longer using the modules in the add_cols
directory to add the repetitive column names, and instead adding
them directly into the database creation script.

Due to the add_cols modules no longer being used, the directory was
deleted from the project.